refactor: make fga package importable
On-behalf-of: @SAP <simon@simontesar.com>
