+++ internal/subroutine/workspace_initializer.go
@@ -15,7 +15,7 @@ import (
 	iclient "github.com/platform-mesh/security-operator/internal/client"
 	logicalclusterclient "github.com/platform-mesh/security-operator/internal/client"
 	"github.com/platform-mesh/security-operator/internal/config"
-	"github.com/platform-mesh/security-operator/internal/fga"
+	"github.com/platform-mesh/security-operator/pkg/fga"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
